pyproject.toml
[project]
name = "agent-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.11",
    "watchdog>=6.0.0",
]
